% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{binnit}
\alias{binnit}
\title{create a bin}
\usage{
binnit(x, y, cuts = seq(0, 1, 0.05), g = 10, pos = "dodge")
}
\arguments{
\item{x}{predictions}

\item{y}{reference}
}
\value{
a ggplot
}
\description{
create a bin
}
