% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_choice}
\alias{assert_choice}
\alias{assert_environment}
\title{Assert specific values and set memberships}
\usage{
assert_choice(x, choices, ...)

assert_environment(x, ...)
}
\arguments{
\item{x}{The variable to assert}

\item{choices}{A vector of values representing the which x must be an
element of.}

\item{...}{Additional parameters passed to corresponding \link{checkmate}
functions \code{\link[checkmate:qassert]{checkmate::qtest()}}, \code{\link[checkmate:checkFlag]{checkmate::check_flag()}}, etc.}
}
\value{
The original object if the assertion passes.
}
\description{
Assert specific values and set memberships
}
