% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{convert_names}
\alias{convert_names}
\title{Take care of variabe names and try to make them consistent}
\usage{
convert_names(x)
}
\arguments{
\item{x}{vector of character names}
}
\value{
converted namesl
}
\description{
Take care of variabe names and try to make them consistent
}
