% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{create_pred_table}
\alias{create_pred_table}
\title{create a 3-group class. receives a data frame with obs prob columns}
\usage{
create_pred_table(DF, thl, thh, levs = c("low", "mid", "high"))
}
\description{
create a 3-group class. receives a data frame with obs prob columns
}
