% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yencode.R
\name{.yencode_map}
\alias{.yencode_map}
\title{Compile an encoding table for yencode}
\usage{
.yencode_map(escape = "\%", whitelist = c("._~-", "][!$&'()*+,;=:/?@#"))
}
\arguments{
\item{escape}{The escape character to use.}

\item{whitelist}{Any characters that should not be escaped.}
}
\value{
A list with the byte lookup table and the multi-byte whitelist
entries that need restoring after the byte pass.
}
\description{
Compile an encoding table for yencode
}
\keyword{internal}
