% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_all_with_prefix.R
\name{import_assertion_functions}
\alias{import_assertion_functions}
\title{Generate imports for all assertions functions}
\usage{
import_assertion_functions(prefix = "assert_", width = 80)
}
\arguments{
\item{prefix}{Prefix used to select exported names.}

\item{width}{Maximum output line width.}
}
\value{
A character string containing \verb{@importFrom} statements and \code{NULL}.
}
\description{
Generates roxygen2 \verb{@importFrom} statements for all assertion functions.
}
\keyword{internal}
