% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_light.R
\name{ll_val_labels}
\alias{ll_val_labels}
\alias{ll_val_labels<-}
\title{Get or set value labels of a labelled variable}
\usage{
ll_val_labels(x, always = FALSE)

ll_val_labels(x) <- value
}
\arguments{
\item{x}{A labelled variable}

\item{always}{Always return at least an empty vector of the correct
type, even if the attribute is not set.}
}
\description{
Gets or sets the value labels (\code{labels} attribute) of a labelled vector. The
getters/setters should be used rather than manipulating attributes directly,
since these functions perform checks to ensure that the result, and the
resulting labelled variable, are valid.
}
\seealso{
Other labelled light:
\code{\link[=ll_chk_labelled]{ll_chk_labelled()}},
\code{\link[=ll_labelled]{ll_labelled()}},
\code{\link[=ll_to_character]{ll_to_character()}},
\code{\link[=ll_var_label]{ll_var_label()}},
\code{\link[=threadbare]{threadbare()}}
}
\concept{labelled light}
\keyword{internal}
