% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_light.R
\name{ll_var_label}
\alias{ll_var_label}
\alias{ll_var_label<-}
\title{Get or set variable label of a labelled variable}
\usage{
ll_var_label(x)

ll_var_label(x) <- value
}
\arguments{
\item{x}{A labelled variable}
}
\description{
Gets or sets the variable label (\code{label} attribute) of a labelled vector. The
getters/setters should be used rather than manipulating attributes directly,
since these functions perform checks to ensure that the result, and the
resulting labelled variable, are valid.
}
\seealso{
Other labelled light:
\code{\link[=ll_chk_labelled]{ll_chk_labelled()}},
\code{\link[=ll_labelled]{ll_labelled()}},
\code{\link[=ll_to_character]{ll_to_character()}},
\code{\link[=ll_val_labels]{ll_val_labels()}},
\code{\link[=threadbare]{threadbare()}}
}
\concept{labelled light}
\keyword{internal}
