% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{n_percent_format}
\alias{n_percent_format}
\title{format text to be used in Rmd's mostly}
\usage{
n_percent_format(tabl, event = "1")
}
\arguments{
\item{event}{the way events are coded}

\item{a}{matrix with observations and events}
}
\value{
converted namesl
}
\description{
format text to be used in Rmd's mostly
}
