% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_tilde.R
\name{recode_tilde}
\alias{recode_tilde}
\title{Recode values using tilde syntax}
\usage{
recode_tilde(x, ..., .default = x, .ptype = NULL)
}
\arguments{
\item{x}{A vector to recode.}

\item{...}{Formulae specifying recoding rules, recoding from \code{lhs} to \code{rhs}.}

\item{.default}{Default value for unmatched inputs. Defaults to \code{x}.}

\item{.ptype}{Optional output type, defaults to \code{.default}. All values in
\code{rhs} and \code{.default} must be mutually compatible with \code{.ptype}}
}
\value{
A vector with recoded values.
}
\description{
Recode elements of a vector using a series of formulas (\code{lhs ~ rhs}) passed
via \code{...}. Each \code{lhs} is matched against elements of \code{x}, and the
corresponding \code{rhs} provides the new value.

This function is closely based on \code{\link[dplyr:case_match]{dplyr::case_match()}} with minimal changes
to make it more intuitive for re-coding tasks. In particular, rather than
setting unmatched values to \code{NA} by default, they remain unchanged
\code{.default}, which itself defaults to \code{x}. The output type can be controlled
with \code{.ptype}. \code{.ptype} defaults to \code{.default}, which means that type can be
changed by setting \code{.default} to either \code{NA} or to a value of the same type
as the \code{rhs} formula values. Incompatibility between the \code{rhs} values and the
\code{.ptype} results in a type error.
}
\examples{
recode_tilde(letters, "a" ~ "first", "z" ~ "last")
recode_tilde(1:5, 1 ~ 10, 2 ~ 20)
# Recoding to different type requires explicit .default values
recode_tilde(1:4, 1 ~ "low", 2 ~ "medium", 3 ~ "high", .default = NA)

}
