% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{glue}
\alias{glue_data}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{glue}{\code{\link[glue:glue]{glue()}}, \code{\link[glue:glue_data]{glue_data()}}}
}}

