% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{retablef}
\alias{retablef}
\title{Format regression results using a table}
\usage{
retablef(obj)
}
\arguments{
\item{obj}{object create by glm() or coxph()}
}
\description{
Format regression results using a table
}
