% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chars.R
\name{sample_chars}
\alias{sample_chars}
\title{Produce a random string}
\usage{
sample_chars(char_num = 20)
}
\arguments{
\item{char_num}{Number of characters in the string.}
}
\value{
A character vector.
}
\description{
Produce a random string
}
