% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{standardize_lookup_table}
\alias{standardize_lookup_table}
\title{Helper function to standardize the \code{lookup_table}.}
\usage{
standardize_lookup_table(lookup_table)
}
\arguments{
\item{lookup_table}{The unstandardized lookup table (must still be one of the
formats specified for the \code{lookup()} function).}
}
\value{
The lookup table as a list.
}
\description{
Preprocessing the lookup table to convert it to a list can take some time, so
when possible, we want to do it only once. Therefore we offload it to a
helper function
}
\keyword{internal}
