% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{transp}
\alias{transp}
\title{like t() but makes sure a data.table is a dataframe and that row names are inserted}
\usage{
transp(DF)
}
\description{
like t() but makes sure a data.table is a dataframe and that row names are inserted
}
