% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uid_anonymise.R
\name{uid_anonymise}
\alias{uid_anonymise}
\title{Anonymise a dataframe}
\usage{
uid_anonymise(.data, id_vars = NULL, ...)
}
\arguments{
\item{.data}{A dataframe.}

\item{id_vars}{Column names containing identifiable information.}

\item{...}{seed_num, char_num, prefix parameters from \link[zmisc]{uid_generate}}
}
\value{
A list of two dataframes.
}
\description{
Anonymise a dataframe
}
\examples{
d <- data.frame(x = letters[1:3], y = 1:3, z = 4:6)
d1 <- uid_anonymise(d, id_vars = c("x"), char_num = 6)
uid_get_key(d1)
uid_get_data(d1)
}
