% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uid_generate.R
\name{uid_generate}
\alias{uid_generate}
\title{Generate random unique ID}
\usage{
uid_generate(n, seed_num = 1, char_num = 20, prefix = NULL)
}
\arguments{
\item{n}{Number of unique IDs to be generated.}

\item{seed_num}{Random seed number for reproducibility.}

\item{char_num}{Number of characters in each unique ID.}

\item{prefix}{Constant prefix to be added to each unique ID.}
}
\value{
A character vector.
}
\description{
Generate random unique ID
}
\examples{
uid_generate(3, seed_num = 1, char_num = 6)
uid_generate(3, seed_num = 99, char_num = 6)
uid_generate(3, seed_num = 99, char_num = 6, prefix = "PK")
}
