% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uid_get_key.R
\name{uid_get_key}
\alias{uid_get_key}
\title{Get dataframe with unique_id and variables for identifiable information}
\usage{
uid_get_key(uid_list)
}
\arguments{
\item{uid_list}{A list created by \link[zmisc]{uid_anonymise}}
}
\value{
A dataframe.
}
\description{
Get dataframe with unique_id and variables for identifiable information
}
\examples{
d <- data.frame(x = letters[1:3], y = 1:3, z = 4:6)
d1 <- uid_anonymise(d, id_vars = c("x"), char_num = 6)
uid_get_key(d1)
uid_get_data(d1)
}
