% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc-utils.R
\name{wrap_error}
\alias{wrap_error}
\title{Utility function to output an error}
\usage{
wrap_error(e, wrap = 50)
}
\arguments{
\item{e}{The error to wrap}

\item{wrap}{How many characters before wrapping}
}
\value{
The error is returned invisibly
}
\description{
This function is used to output errors in examples in a tidy manner.
}
\keyword{internal}
