% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zingle.R
\name{zingle}
\alias{zingle}
\title{Return the single (unique) value found in a vector}
\usage{
zingle(
  x,
  ...,
  empty.ok = FALSE,
  na.ok.partial = FALSE,
  na.ok.all = FALSE,
  nan.ok.all = FALSE
)
}
\arguments{
\item{x}{Vector of elements that should all be identical}

\item{...}{Unused, reserved to force later arguments to be named}

\item{empty.ok}{Is an empty vector ok?}

\item{na.ok.partial}{Is a mix of \code{NA} and one distinct non-missing value ok?}

\item{na.ok.all}{Is a vector of only \code{NA} values ok?}

\item{nan.ok.all}{Is \code{NaN} ok as the returned value?}
}
\value{
The single element in the vector, unnamed. For \code{na.ok.partial} the
non-missing value is returned.
}
\description{
\code{\link[=zingle]{zingle()}} returns the only value present in a vector. If the vector contains
more than one distinct value, it throws an error. This is a guard for
aggregations where all values within a group should be identical, but where
you want that assumption checked rather than assumed. Only values are
compared. Names are ignored and the result is unnamed, in line with other
aggregation functions.
}
\examples{
zingle(c("Alpha", "Alpha", "Alpha"))
zingle(c("Alpha", NA, "Alpha"), na.ok.partial = TRUE)
zingle(c(NA, NA), na.ok.all = TRUE)
zingle(c(NaN, NaN), nan.ok.all = TRUE)

}
