% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{chk_dots_empty}
\alias{chk_dots_empty}
\title{Assert that no dots arguments are passed}
\usage{
chk_dots_empty(
  env = caller_env(),
  error = NULL,
  call = caller_env(),
  action = abort
)
}
\arguments{
\item{env}{Environment in which to look for \code{...}.}

\item{error}{An optional error handler passed to \code{\link[rlang:try_fetch]{try_fetch()}}. Use
this e.g. to demote an error into a warning.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{action}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\description{
\code{chk_dots_empty()} is an alias for \code{\link[rlang:check_dots_empty]{rlang::check_dots_empty()}}, provided
for naming consistency with other assertion functions. It throws an error if
any arguments are passed through \code{...}.
}
