% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue_vector}
\alias{glue_vector}
\title{Glue interpolation vectors in pipes}
\usage{
glue_vector(
  .,
  template = "{.}",
  ...,
  .sep = "",
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  .na = "NA",
  .null = character(),
  .comment = "#",
  .literal = FALSE,
  .transformer = glue::identity_transformer,
  .trim = TRUE
)
}
\arguments{
\item{.}{A character vector to be interpolated.}

\item{template}{A glue template string. Use \code{{.}} to refer to the default
(unnamed) vector variable, or the names of any other variables accessible
in the relevant environment. Variables are recycled using tidyverse
recycling rules.}

\item{...}{Reserved and should not be used.}

\item{.sep, .envir, .open, .close, .na, .null, .comment, .literal, .transformer, .trim}{Arguments passed on to \code{\link[glue:glue]{glue::glue()}}. Must be passed by name.
See \code{\link[glue:glue]{glue::glue()}} for details.}
}
\value{
A character vector with interpolated values. The length is determined
by tidyverse recycling rules for all referenced variables.
}
\description{
Applies \code{glue::glue()} to each element of a character vector using a template
string, enabling pipe-friendly, element-wise interpolation. Useful when the
vector to process is not encapsulated in a \code{data.frame} or other
environment-like object.
}
\examples{
  letters |> glue_vector("Letters include {.} and {LETTERS}")

}
