% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_light.R
\name{ll_assert_labelled}
\alias{ll_assert_labelled}
\title{Verify that x is a valid labelled variable}
\usage{
ll_assert_labelled(x)
}
\arguments{
\item{x}{A labelled variable}
}
\value{
Invisibly returns x if the check is successful.
}
\description{
Verify that x is a valid labelled variable satisfying the (minimal)
specification inherent in the parameter documentation of the
\code{\link[haven:labelled]{haven::labelled()}} function for \code{haven_labelled} objects.
}
\seealso{
Other labelled light: 
\code{\link{ll_labelled}()},
\code{\link{ll_to_character}()},
\code{\link{ll_val_labels}()},
\code{\link{ll_var_label}()},
\code{\link{threadbare}()}
}
\concept{labelled light}
\keyword{internal}
