% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yencode.R
\name{yencode}
\alias{yencode}
\alias{yencoder}
\alias{ydecode}
\alias{ydecoder}
\title{Yet (another urlencode compatible) encoding scheme}
\usage{
yencode(string, escape = "\%", whitelist = c("._~-", "][!$&'()*+,;=:/?@#"))

yencoder(escape = "\%", whitelist = c("._~-", "][!$&'()*+,;=:/?@#"))

ydecode(string, escape = "\%")

ydecoder(escape = "\%")
}
\arguments{
\item{string}{The string to process.}

\item{escape}{The escape character to use.}

\item{whitelist}{Any characters that should not be escaped. See details.}
}
\value{
The processed (encoded or decoded) string.
}
\description{
Yet (another urlencode compatible) encoding scheme
}
