% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc-package.R
\docType{package}
\name{zmisc-package}
\alias{zmisc}
\alias{zmisc-package}
\title{zmisc: Vector Look-Ups and Safer Sampling}
\description{
A collection of utility functions that facilitate looking up vector values from a lookup table, annotate values in a table for clearer viewing, and support a safer approach to vector sampling, sequence generation, and aggregation. Also included is a family of argument checks which return their input so that they compose nicely in a pipe.
}
\details{
For an overview with examples, see the package website at
\url{https://torfason.github.io/zmisc/}. For the argument checks, see
\code{vignette("chk")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/torfason/zmisc/}
  \item \url{https://torfason.github.io/zmisc/}
}

}
\keyword{internal}
