% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc-package.R
\docType{package}
\name{zmisc}
\alias{zmisc}
\alias{zmisc-package}
\title{zmisc: Vector Look-Ups and Safer Sampling}
\description{
A collection of utility functions that facilitate
looking up vector values from a lookup  table, and support a
safer approach to vector  sampling, sequence generation,
and aggregation.

For more information, see vignette("zmisc").
}
\keyword{internal}
