% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc-package.R
\docType{package}
\name{zmisc}
\alias{zmisc}
\alias{zmisc-package}
\title{zmisc: Vector Lookups and Safer Sampling}
\description{
This package contains several utility functions,
including functions to look up vector values from a lookup
table, and functions that support a safer approach to vector
sampling, sequence generation, and aggregation.

For more information, see vignette("zmisc").
}
\keyword{internal}
